fix(ntp): forzar conversión a int para evitar error de tipo en formato string
- Agregado int() en ntp.py línea 11 para tz_offset
- Agregado int() en timezone.py línea 33 para valor de retorno
- MicroPython round() puede retornar float, causando TypeError al usar :+d
- Asegura que tz_offset sea int antes de formatear como {tz_offset:+d}

+++ src/timezone.py
@@ -30,7 +30,7 @@ def get_timezone_offset(longitude):
     """
     # Timezone = longitude / 15 degrees per hour
     # Round to nearest hour for simplicity
-    return round(longitude / 15.0)
+    return int(round(longitude / 15.0))
 
 def apply_timezone_to_time(time_tuple, tz_offset):
     """
